mejoras del codigo SpringBatch

+++ src/main/java/com/batch/SpringBatchApplication/dto/ProductoDTO.java
@@ -9,5 +9,5 @@ public class ProductoDTO {
 
     private Long id;
     private String nombre;
-    private BigDecimal precio;
+    private Float precio;
 }
