+++ src/main/java/com/batch/SpringBatchApplication/model/Producto.java
@@ -13,5 +13,5 @@ public class Producto {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private String nombre;
-    private BigDecimal precio;
+    private Float precio;
 }
